OSGI_JAX-RS
This project aims to create a tool to manage OSGi bundles deployed on an
EQUINOX container. we can find there the interface "BundleUtils" which
contains the methods may be called by the Simple client. It contains
also "BundleUtilsImpl"; the implementation of methods of
"BundlesUtils"(getBundles,installBundle,startBundle...) and also the
"ServiceActivator" used for the registration of the service.